aws-sms/requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
moto[kinesis,dynamodb,s3,sns,pinpoint]>=5.0.0

# Code quality
ruff>=0.8.0
mypy>=1.13.0
black>=24.0.0

# Type stubs
boto3-stubs[kinesis,dynamodb,s3,sns,pinpoint]>=1.35.0
